+++ CHANGELOG.md
@@ -79,6 +79,8 @@ being instantiated.
   `Krizalys\Onedrive\Proxy\DriveItemProxy::invite()`.
 - `Krizalys\Onedrive\Proxy\HashesProxy`.
 - `Krizalys\Onedrive\Proxy\FolderViewProxy`.
+- Support for creating shared folders: see
+  `Krizalys\Onedrive\Proxy\DriveProxy::createSharedFolder()`.
 
 ### Changed